sql/export: extract EXPORT files from sql/importer
There is no point in keeping EXPORT-related files in the same package as
IMPORT-related ones. This is beneficial also due to different owners.
The only downside is that backports for EXPORT changes won't be clean
for a couple of releases.

I copied using "large" pool for "heavy" test configs (i.e. duress).

Release note: None
