+++ pkg/BUILD.bazel
@@ -482,6 +482,7 @@ ALL_TESTS = [
     "//pkg/sql/execinfrapb:execinfrapb_test",
     "//pkg/sql/execstats:execstats_disallowed_imports_test",
     "//pkg/sql/execstats:execstats_test",
+    "//pkg/sql/export:export_test",
     "//pkg/sql/exprutil:exprutil_test",
     "//pkg/sql/flowinfra:flowinfra_disallowed_imports_test",
     "//pkg/sql/flowinfra:flowinfra_test",
@@ -1990,6 +1991,8 @@ GO_TARGETS = [
     "//pkg/sql/execstats:execstats",
     "//pkg/sql/execstats:execstats_test",
     "//pkg/sql/execversion:execversion",
+    "//pkg/sql/export:export",
+    "//pkg/sql/export:export_test",
     "//pkg/sql/exprutil:evalexpr",
     "//pkg/sql/exprutil:exprutil",
     "//pkg/sql/exprutil:exprutil_test",
