+++ pkg/sql/export/BUILD.bazel
@@ -1,0 +1,89 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "export",
+    srcs = [
+        "export_base.go",
+        "exportcsv.go",
+        "exportparquet.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/export",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/cloud",
+        "//pkg/roachpb",
+        "//pkg/settings",
+        "//pkg/sql/catalog/colinfo",
+        "//pkg/sql/execinfra",
+        "//pkg/sql/execinfrapb",
+        "//pkg/sql/pgwire/pgcode",
+        "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/rowenc",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/types",
+        "//pkg/util/encoding/csv",
+        "//pkg/util/mon",
+        "//pkg/util/parquet",
+        "//pkg/util/tracing",
+        "//pkg/util/unique",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
+
+go_test(
+    name = "export_test",
+    srcs = [
+        "exportcsv_test.go",
+        "exportparquet_test.go",
+        "main_test.go",
+    ],
+    exec_properties = select({
+        "//build/toolchains:is_heavy": {"test.Pool": "large"},
+        "//conditions:default": {"test.Pool": "default"},
+    }),
+    deps = [
+        ":export",
+        "//pkg/base",
+        "//pkg/ccl",
+        "//pkg/col/coldata",
+        "//pkg/config",
+        "//pkg/config/zonepb",
+        "//pkg/kv/kvpb",
+        "//pkg/kv/kvserver",
+        "//pkg/multitenant/tenantcapabilitiespb",
+        "//pkg/security/securityassets",
+        "//pkg/security/securitytest",
+        "//pkg/security/username",
+        "//pkg/server",
+        "//pkg/settings/cluster",
+        "//pkg/sql",
+        "//pkg/sql/catalog/colinfo",
+        "//pkg/sql/catalog/descidgen",
+        "//pkg/sql/execinfra",
+        "//pkg/sql/execinfrapb",
+        "//pkg/sql/isql",
+        "//pkg/sql/randgen",
+        "//pkg/sql/rowenc",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/sessiondata",
+        "//pkg/sql/types",
+        "//pkg/testutils",
+        "//pkg/testutils/serverutils",
+        "//pkg/testutils/skip",
+        "//pkg/testutils/sqlutils",
+        "//pkg/testutils/testcluster",
+        "//pkg/util/ctxgroup",
+        "//pkg/util/hlc",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
+        "//pkg/util/mon",
+        "//pkg/util/parquet",
+        "//pkg/util/randutil",
+        "//pkg/util/syncutil",
+        "//pkg/workload/bank",
+        "//pkg/workload/workloadsql",
+        "@com_github_gogo_protobuf//proto",
+        "@com_github_jackc_pgx_v5//:pgx",
+        "@com_github_stretchr_testify//require",
+    ],
+)
