+++ pkg/sql/export/exportcsv_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package importer_test
+package export_test
 
 import (
 	"bytes"
