+++ pkg/sql/importer/BUILD.bazel
@@ -9,9 +9,6 @@ filegroup(
 go_library(
     name = "importer",
     srcs = [
-        "export_base.go",
-        "exportcsv.go",
-        "exportparquet.go",
         "import_job.go",
         "import_planning.go",
         "import_processor.go",
@@ -105,8 +102,6 @@ go_library(
         "//pkg/util/log",
         "//pkg/util/log/eventpb",
         "//pkg/util/log/logutil",
-        "//pkg/util/mon",
-        "//pkg/util/parquet",
         "//pkg/util/protoutil",
         "//pkg/util/retry",
         "//pkg/util/syncutil",
@@ -114,7 +109,6 @@ go_library(
         "//pkg/util/timeutil",
         "//pkg/util/timeutil/pgdate",
         "//pkg/util/tracing",
-        "//pkg/util/unique",
         "//pkg/workload",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
@@ -135,8 +129,6 @@ go_test(
         "client_import_test.go",
         "csv_internal_test.go",
         "csv_testdata_helpers_test.go",
-        "exportcsv_test.go",
-        "exportparquet_test.go",
         "import_csv_mark_redaction_test.go",
         "import_into_test.go",
         "import_mvcc_test.go",
@@ -164,15 +156,11 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/blobs",
-        "//pkg/ccl",
         "//pkg/cloud",
         "//pkg/cloud/cloudpb",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/cloud/nodelocal",
         "//pkg/cloud/userfile",
-        "//pkg/col/coldata",
-        "//pkg/config",
-        "//pkg/config/zonepb",
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/jobs/jobstest",
@@ -184,7 +172,6 @@ go_test(
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/kv/kvserver/protectedts",
-        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
@@ -196,9 +183,7 @@ go_test(
         "//pkg/sql/catalog/bootstrap",
         "//pkg/sql/catalog/catformat",
         "//pkg/sql/catalog/catpb",
-        "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/dbdesc",
-        "//pkg/sql/catalog/descidgen",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/tabledesc",
@@ -240,8 +225,6 @@ go_test(
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/log/eventpb",
-        "//pkg/util/mon",
-        "//pkg/util/parquet",
         "//pkg/util/protoutil",
         "//pkg/util/randutil",
         "//pkg/util/retry",
@@ -250,16 +233,12 @@ go_test(
         "//pkg/util/timeutil/pgdate",
         "//pkg/util/tracing",
         "//pkg/workload",
-        "//pkg/workload/bank",
         "//pkg/workload/tpcc",
-        "//pkg/workload/workloadsql",
         "@com_github_cockroachdb_cockroach_go_v2//crdb",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_go_sql_driver_mysql//:mysql",
-        "@com_github_gogo_protobuf//proto",
         "@com_github_jackc_pgconn//:pgconn",
         "@com_github_jackc_pgx_v4//:pgx",
-        "@com_github_jackc_pgx_v5//:pgx",
         "@com_github_kr_pretty//:pretty",
         "@com_github_lib_pq//:pq",
         "@com_github_linkedin_goavro_v2//:goavro",
